+++ module/extension/model.php
@@ -3,7 +3,7 @@
  * The model file of extension module of ZenTaoCMS.
  *
  * @copyright   Copyright 2009-2015 青岛易软天创网络科技有限公司(QingDao Nature Easy Soft Network Technology Co,LTD, www.cnezsoft.com)
- * @license     ZPL (http://zpl.pub/page/zplv11.html)
+ * @license     ZPL (http://zpl.pub/page/zplv12.html)
  * @author      Chunsheng Wang <chunsheng@cnezsoft.com>
  * @package     extension
  * @version     $Id$
@@ -15,14 +15,6 @@ class extensionModel extends model
      * The extension manager version. Don't change it. 
      */
     const EXT_MANAGER_VERSION = '1.3';
-
-    /**
-     * The api agent(use snoopy).
-     * 
-     * @var object   
-     * @access public
-     */
-    public $agent;
 
     /**
      * The api root.
@@ -41,20 +33,8 @@ class extensionModel extends model
     public function __construct()
     {
         parent::__construct();
-        $this->setAgent();
         $this->setApiRoot();
         $this->classFile = $this->app->loadClass('zfile');
-    }
-
-    /**
-     * Set the api agent.
-     * 
-     * @access public
-     * @return void
-     */
-    public function setAgent()
-    {
-        $this->agent = $this->app->loadClass('snoopy');
     }
 
     /**
@@ -77,9 +57,23 @@ class extensionModel extends model
      */
     public function fetchAPI($url)
     {
-        $url .= (strpos($url, '?') === false ? '?' : '&') . 'lang=' . str_replace('-', '_', $this->app->getClientLang()) . '&managerVersion=' . self::EXT_MANAGER_VERSION . '&zentaoVersion=' . $this->config->version;
-        $this->agent->fetch($url);
-        $result = json_decode($this->agent->results);
+        $this->app->loadConfig('upgrade');
+        $version = $this->config->version;
+        if(strpos($version, 'biz') !== false)
+        {
+            $version = str_replace('.', '_', $version);
+            if(isset($this->config->bizVersion[$version])) $version = $this->config->bizVersion[$version];
+            $version = str_replace('_', '.', $version);
+        }
+        if(strpos($version, 'pro') !== false)
+        {
+            $version = str_replace('.', '_', $version);
+            if(isset($this->config->proVersion[$version])) $version = $this->config->proVersion[$version];
+            $version = str_replace('_', '.', $version);
+        }
+
+        $url .= (strpos($url, '?') === false ? '?' : '&') . 'lang=' . str_replace('-', '_', $this->app->getClientLang()) . '&managerVersion=' . self::EXT_MANAGER_VERSION . '&zentaoVersion=' . $version;
+        $result = json_decode(common::http($url));
 
         if(!isset($result->status)) return false;
         if($result->status != 'success') return false;
@@ -170,8 +164,7 @@ class extensionModel extends model
     public function downloadPackage($extension, $downLink)
     {
         $packageFile = $this->getPackageFile($extension);
-        $this->agent->fetch($downLink);
-        file_put_contents($packageFile, $this->agent->results);
+        file_put_contents($packageFile, common::http($downLink));
     }
 
     /**
@@ -183,7 +176,7 @@ class extensionModel extends model
      */
     public function getLocalExtensions($status)
     {
-        $extensions = $this->dao->select('*')->from(TABLE_EXTENSION)->where('status')->eq($status)->fi()->fetchAll('code');
+        $extensions = $this->dao->select('*')->from(TABLE_EXTENSION)->where('status')->in($status)->fi()->fetchAll('code');
         foreach($extensions as $extension)
         {
             if($extension->site and stripos(strtolower($extension->site), 'http') === false) $extension->site = 'http://' . $extension->site;
@@ -264,8 +257,12 @@ class extensionModel extends model
         $info = (object)spyc_load(file_get_contents($infoFile));
         if(isset($info->releases))
         {
-            krsort($info->releases);
             $info->version = key($info->releases);
+            foreach(array_keys($info->releases) as $version)
+            {
+                if(version_compare($info->version, $version, '<')) $info->version = $version;
+            }
+
             foreach($info->releases[$info->version] as $key => $value) $info->$key = $value;
         }
         return $info;
@@ -674,7 +671,7 @@ class extensionModel extends model
 
         /* Remove the extracted files. */
         $extractedDir = realpath("ext/$extension");
-        if($extractedDir != '/' and !$this->classFile->removeDir($extractedDir))
+        if($extractedDir and $extractedDir != '/' and !$this->classFile->removeDir($extractedDir))
         {
             $removeCommands[] = PHP_OS == 'Linux' ? "rm -fr $extractedDir" : "rmdir $extractedDir /s";
         }
@@ -881,5 +878,36 @@ class extensionModel extends model
 
         if($type != 'between') return !$result;
         return $result;
+    }
+
+    /**
+     * Get extension expire date.
+     * 
+     * @param  int    $extension 
+     * @access public
+     * @return void
+     */
+    public function getExpireDate($extension)
+    {
+        $licencePath = $this->app->getConfigRoot() . 'license/';
+        $today       = date('Y-m-d');
+        $expireDate  = '';
+
+        $licenceOrderFile = $licencePath . 'order' . $extension->code . $extension->version . '.txt';
+        if(file_exists($licenceOrderFile))
+        {
+            $order = file_get_contents($licenceOrderFile);
+            $order = unserialize($order);
+            if($order->type != 'life')
+            {
+                $days = isset($order->days) ? $order->days : 0;
+                if($order->type == 'demo') $days = 31;
+                if($order->type == 'year') $days = 365;
+                $startDate  = $order->paidDate != '0000-00-00 00:00:00' ? $order->paidDate : $order->createdDate;
+                if($days) $expireDate = date('Y-m-d', strtotime($startDate) + $days * 24 * 3600);
+            }
+        }
+
+        return $expireDate;
     }
 }
